requirements.txt
numpy>=1.26,<2
Pillow>=10,<13
requests>=2.31,<3
beautifulsoup4>=4.12,<5
huggingface_hub>=0.30
transformers>=5,<6
accelerate>=1.13,<2
paddlepaddle>=3.0,<4
paddleocr>=2.8,<4
ultralytics>=8.3,<9
clip @ git+https://github.com/ultralytics/CLIP.git
sentencepiece>=0.1.99,<1.0
fastapi>=0.111,<1
uvicorn>=0.30,<1
jinja2>=3.1,<4
jsonschema>=4.23,<5
